chore: changed lib name
