+++ packages/vue-components/nb-payments/src/components/__tests__/NbCreditCard.spec.js
@@ -1,0 +1,11 @@
+import { describe, it, expect } from 'vitest';
+
+import { mount } from '@vue/test-utils';
+import NbCreditCard from '../NbCreditCard.vue';
+
+describe('NbCreditCard', () => {
+  it('renders properly', () => {
+    const wrapper = mount(NbCreditCard, { props: { txt: 'Hello Vitest' } });
+    expect(wrapper.text()).toContain('Hello Vitest');
+  });
+});
