+++ packages/vue-components/nb-payments/src/plugin/masterTool.js
@@ -1,0 +1,11 @@
+import MasterTool from '@vlalg-nimbus/master-tool'
+
+const plugin = {
+  install(Vue) {
+    const $masterTool = MasterTool()
+
+    Vue.provide('$masterTool', $masterTool)
+  },
+};
+
+export default plugin;
